refactor scan logic
